requirements.txt
appdirs==1.4.4
asgiref==3.3.1
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2020.12.5
chardet==4.0.0
cryptography==2.1.4
cupshelpers==1.0
distlib==0.3.1
Django==3.0
django-appconf==1.0.4
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-phone-field==1.8.1
django-rest-framework==0.1.0
django-upload-validator==1.1.5
djangorestframework==3.12.2
djangorestframework-files==1.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
keyring==10.6.0
keyrings.alt==3.0
louis==3.5.0
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==2.0.3
netifaces==0.10.4
olefile==0.45.1
packaging==20.9
pexpect==4.2.1
pilkit==2.0
Pillow==8.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pyparsing==2.4.7
python-apt==1.6.5+ubuntu0.5
python-debian==0.1.32
python-dotenv==0.15.0
python-magic==0.4.22
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.25.1
SecretStorage==2.3.1
six==1.15.0
sqlparse==0.4.1
style==1.1.0
toml==0.10.2
tox==3.23.0
ufw==0.36
update==0.0.1
urllib3==1.26.4
usb-creator==0.3.3
virtualenv==20.0.17
